python/pyproject.toml
[project]
name = "srs-rl"
version = "0.0.1"
description = "Reinforcement learning library for the SRS simulator"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "grpcio>=1.75.1",
    "grpcio-tools>=1.75.1",
    "numpy>=1.27.0",
    "matplotlib>=3.8.0",
    "gymnasium>=0.30.0",
    "tqdm>=4.67.1",
    "tensorflow>=2.20.0",
    "imageio>=2.31.4",
    "pygame>=2.6.1",
    "notebook>=7.4.7",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.10",
]

[tool.uv]
managed = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.0",
    "ty>=0.0.1a22",
]

[tool.ruff]
exclude = ["*_pb2*.py", "*.ipynb"]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B", "A", "COM"]
ignore = ["E501", "COM812", "S311", "S603", "S607"]
extend-select = ["C4", "T20", "S", "RET", "TCH"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # Allow assert in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

